pyproject.toml
[project]
name = "autoagent-vnext"
version = "0.1.0"
description = "Self-healing, self-optimizing ADK agent following Karpathy's autoresearch pattern"
requires-python = ">=3.11"
dependencies = [
    "google-adk>=0.3.0",
    "fastapi>=0.110.0",
    "uvicorn>=0.29.0",
    "click>=8.1.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
]

[project.scripts]
autoagent = "runner:cli"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
